hasMobile-main/www/flaskapp/requirements.txt
click
Flask
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python
protobuf
python-slugify
six
text-unidecode
timeago
Werkzeug
flask_sitemapper
flask-sitemapper
cryptography
geoip2
requests
Flask-APScheduler
selenium
bs4
pymongo